Extract parsing functions into a separate module
Refactored nlp4attributes by removing all attribute parsing functions and placing them into a separate file attribute_extractor.py. This better helps modularize the attribute extraction code and achieve better separation of concerns.

From the perspective of the backend API, nlp4attributes is supposed to receive a pdf file name and print out a json object containing the document's attributes. As such, nlp4attributes is effectively a script that receives some input, passes it to the relevant functions, and returns the output. As a module, it shouldn't need to know how to parse these documents, rather that logic can live elsewhere.
